split macro helpers in seperate modules
